# schemas/openapi/wp16-openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The one error shape this API returns.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code.",
            "maxLength": 64,
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails",
            "description": "Bounded machine-readable metadata."
          },
          "message": {
            "description": "Controlled message from a fixed catalogue. Never the text of an unexpected exception.",
            "maxLength": 512,
            "type": "string"
          },
          "request_id": {
            "description": "The same identifier the response header carries.",
            "format": "uuid",
            "maxLength": 36,
            "type": "string"
          }
        },
        "required": [
          "code",
          "details",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetails": {
        "additionalProperties": false,
        "description": "Bounded safe metadata about a failure. Carries locations and codes; never a rejected value, a phenotype, a medication list, a path, a query or a stack frame.",
        "properties": {
          "components": {
            "description": "Named components that were not ready.",
            "items": {
              "maxLength": 64,
              "type": "string"
            },
            "maxItems": 20,
            "type": "array"
          },
          "issues": {
            "description": "One entry per contract violation.",
            "items": {
              "$ref": "#/components/schemas/ErrorIssue"
            },
            "maxItems": 20,
            "type": "array"
          },
          "limit": {
            "anyOf": [
              {
                "maximum": 1000000,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The bound a request exceeded."
          },
          "required_role": {
            "anyOf": [
              {
                "enum": [
                  "DEMO_USER",
                  "EXPERT_REVIEWER",
                  "ADMIN"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The role a forbidden request would have needed."
          },
          "work_package": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The work package that will implement a not-implemented route."
          }
        },
        "required": [],
        "title": "ErrorDetails",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Every failure response, for every status, has this shape.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "The failure."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ErrorIssue": {
        "additionalProperties": false,
        "description": "One contract violation. ``location`` is a normalised JSON pointer and ``code`` is stable; neither the supplied value nor any text derived from it appears.",
        "properties": {
          "code": {
            "description": "Stable violation code.",
            "maxLength": 64,
            "type": "string"
          },
          "location": {
            "description": "Normalised JSON pointer, for example $.profile.observations[0].value.",
            "maxLength": 128,
            "type": "string"
          }
        },
        "required": [
          "code",
          "location"
        ],
        "title": "ErrorIssue",
        "type": "object"
      },
...
      "ExpertReviewAssignmentListResponse": {
        "additionalProperties": false,
        "description": "The calling reviewer's assignments and nobody else's. A reviewer who could list another's would learn which cases exist and who holds them.",
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/ExpertReviewAssignmentSummary"
            },
            "maxItems": 20,
            "type": "array"
          },
          "count": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "assignments",
          "count"
        ],
        "title": "ExpertReviewAssignmentListResponse",
        "type": "object"
      },
      "ExpertReviewAssignmentSummary": {
        "additionalProperties": false,
        "description": "One of the calling reviewer's own assignments.",
        "properties": {
          "blinded": {
            "type": "boolean"
          },
          "case_id": {
            "maxLength": 64,
            "type": "string"
          },
          "release_public_id": {
            "maxLength": 64,
            "type": "string"
          },
          "review_id": {
            "maxLength": 64,
            "type": "string"
          },
          "state": {
            "maxLength": 64,
            "type": "string"
          }
        },
        "required": [
          "blinded",
          "case_id",
          "release_public_id",
          "review_id",
          "state"
        ],
        "title": "ExpertReviewAssignmentSummary",
        "type": "object"
      },
...
      "ExpertReviewStateResponse": {
        "additionalProperties": false,
        "description": "The reviewer's own view of their assignment while blinded. There is no result field in this shape - not an optional one, not a nullable one. A hidden value is still disclosure, so the pre-reveal response has nowhere to put one.",
        "properties": {
          "assignment_id": {
            "maxLength": 64,
            "type": "string"
          },
          "blinded": {
            "description": "True until a reveal record exists.",
            "type": "boolean"
          },
          "case_id": {
            "maxLength": 64,
            "type": "string"
          },
          "correction_count": {
            "minimum": 0,
            "type": "integer"
          },
          "expectation_recorded": {
            "type": "boolean"
          },
          "expectation_revision": {
            "anyOf": [
              {
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "expectation_revision_hash": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "protocol_version": {
            "maxLength": 64,
            "type": "string"
          },
          "release_public_id": {
            "maxLength": 64,
            "type": "string"
          },
          "review_id": {
            "maxLength": 64,
            "type": "string"
          },
          "state": {
            "description": "ASSIGNED, EXPECTATION_RECORDED, RESULT_REVEALED, COMPLETED or INVALIDATED.",
            "maxLength": 64,
            "type": "string"
          }
        },
        "required": [
          "assignment_id",
          "blinded",
          "case_id",
          "correction_count",
          "expectation_recorded",
          "protocol_version",
          "release_public_id",
          "review_id",
          "state"
        ],
        "title": "ExpertReviewStateResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "PgxPrincipal": {
        "description": "A stub contract, not an implemented scheme. WP-23 owns authentication; this deployment establishes a principal through a configured provider and refuses with 503 when none is configured. No token is issued by this API and no login endpoint exists.",
        "scheme": "bearer",
        "type": "http",
        "x-pgx-implemented": false,
        "x-pgx-owned-by": "WP-23"
      }
    }
  },
  "info": {
    "description": "Structured pharmacogenomic coverage and attention facts.\n\nThis API reports what governed rules said about the axes that could be evaluated (attention) and how much of the question could be evaluated (coverage). It reports the two together and neither summarises the other.\n\nIt does not diagnose, does not prescribe, does not calculate a dose, does not select or rank a medicine, and does not tell anyone to change a medicine. Every example in this document is synthetic: no example here is a real case, a real assessment, a real evidence record or a real credential.",
    "title": "PGx Platform V2 API",
    "version": "1.0.0",
    "x-pgx-contract-version": "pgx-api-contract/1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/expert-reviews": {
      "get": {
        "description": "Returns only assignments held by the authenticated principal. A reviewer who could list another's would learn which holdout cases exist and who holds them, which is the linkage a blind protocol keeps closed.",
        "operationId": "listExpertReviewAssignments",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertReviewAssignmentListResponse"
                }
              }
            },
            "description": "The calling reviewer's own assignments"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "REQUEST_ID_INVALID",
                    "details": {},
                    "message": "The supplied request identifier is not a canonical UUID.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The supplied request identifier is not a canonical UUID. Codes: REQUEST_ID_INVALID, REQUEST_MALFORMED"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHENTICATED",
                    "details": {},
                    "message": "This endpoint requires an authenticated principal.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "This endpoint requires an authenticated principal. Codes: UNAUTHENTICATED"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_PROTOCOL_NOT_APPROVED",
                    "details": {},
                    "message": "The blind review protocol has not been approved by named human and scientific reviewers.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The blind review protocol has not been approved by named human and scientific reviewers. Codes: EXPERT_REVIEW_PROTOCOL_NOT_APPROVED, EXPERT_REVIEW_ROLE_REQUIRED, FORBIDDEN_ROLE"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_NOT_ASSIGNED",
                    "details": {},
                    "message": "No active assignment. Returned identically whether the case is unknown, is not an expert-holdout case, or belongs to another reviewer.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No active assignment. Returned identically whether the case is unknown, is not an expert-holdout case, or belongs to another reviewer. Codes: EXPERT_REVIEW_NOT_ASSIGNED"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_CASE_MANIFEST_MISMATCH",
                    "details": {},
                    "message": "The case manifest hash no longer matches the pinned one.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The case manifest hash no longer matches the pinned one. Codes: EXPERT_REVIEW_CASE_MANIFEST_MISMATCH, EXPERT_REVIEW_INVALID_TRANSITION, EXPERT_REVIEW_PROTOCOL_MISMATCH, EXPERT_REVIEW_RELEASE_MISMATCH"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_FORGED_FIELD",
                    "details": {},
                    "message": "The request supplied a field the server owns.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request supplied a field the server owns. Codes: EXPERT_REVIEW_FORGED_FIELD"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_AUDIT_FAILED",
                    "details": {},
                    "message": "The audit event could not be appended, so the governed action was rolled back.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The audit event could not be appended, so the governed action was rolled back. Codes: EXPERT_REVIEW_AUDIT_FAILED, INTERNAL_ERROR"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTHENTICATION_NOT_CONFIGURED",
                    "details": {},
                    "message": "No authentication provider is configured.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No authentication provider is configured. Codes: AUTHENTICATION_NOT_CONFIGURED, EXPERT_REVIEW_NOT_AVAILABLE"
          }
        },
        "security": [
          {
            "PgxPrincipal": []
          }
        ],
        "summary": "The calling reviewer's own assignments",
        "tags": [
          "expert-review"
        ],
        "x-pgx-required-roles": [
          "EXPERT_REVIEWER"
        ]
      }
    },
    "/api/v1/expert-reviews/{case_id}": {
      "get": {
        "description": "The response model has no result field before a reveal exists - not an optional one and not a nullable one. A hidden value is still disclosure, so the pre-reveal shape has nowhere to put one. A missing or foreign assignment returns the same refusal as an unknown case.",
        "operationId": "getExpertReviewState",
        "parameters": [
          {
            "description": "An expert-review case identifier. Not resolved in P0.",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertReviewStateResponse"
                }
              }
            },
            "description": "One assignment's state, blinded until reveal"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "REQUEST_ID_INVALID",
                    "details": {},
                    "message": "The supplied request identifier is not a canonical UUID.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The supplied request identifier is not a canonical UUID. Codes: REQUEST_ID_INVALID, REQUEST_MALFORMED"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "UNAUTHENTICATED",
                    "details": {},
                    "message": "This endpoint requires an authenticated principal.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "This endpoint requires an authenticated principal. Codes: UNAUTHENTICATED"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_PROTOCOL_NOT_APPROVED",
                    "details": {},
                    "message": "The blind review protocol has not been approved by named human and scientific reviewers.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The blind review protocol has not been approved by named human and scientific reviewers. Codes: EXPERT_REVIEW_PROTOCOL_NOT_APPROVED, EXPERT_REVIEW_ROLE_REQUIRED, FORBIDDEN_ROLE"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_NOT_ASSIGNED",
                    "details": {},
                    "message": "No active assignment. Returned identically whether the case is unknown, is not an expert-holdout case, or belongs to another reviewer.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No active assignment. Returned identically whether the case is unknown, is not an expert-holdout case, or belongs to another reviewer. Codes: EXPERT_REVIEW_NOT_ASSIGNED"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_CASE_MANIFEST_MISMATCH",
                    "details": {},
                    "message": "The case manifest hash no longer matches the pinned one.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The case manifest hash no longer matches the pinned one. Codes: EXPERT_REVIEW_CASE_MANIFEST_MISMATCH, EXPERT_REVIEW_INVALID_TRANSITION, EXPERT_REVIEW_PROTOCOL_MISMATCH, EXPERT_REVIEW_RELEASE_MISMATCH"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_FORGED_FIELD",
                    "details": {},
                    "message": "The request supplied a field the server owns.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request supplied a field the server owns. Codes: EXPERT_REVIEW_FORGED_FIELD"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "EXPERT_REVIEW_AUDIT_FAILED",
                    "details": {},
                    "message": "The audit event could not be appended, so the governed action was rolled back.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The audit event could not be appended, so the governed action was rolled back. Codes: EXPERT_REVIEW_AUDIT_FAILED, INTERNAL_ERROR"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTHENTICATION_NOT_CONFIGURED",
                    "details": {},
                    "message": "No authentication provider is configured.",
                    "request_id": "00000000-0000-4000-8000-000000000000"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No authentication provider is configured. Codes: AUTHENTICATION_NOT_CONFIGURED, EXPERT_REVIEW_NOT_AVAILABLE"
          }
        },
        "security": [
          {
            "PgxPrincipal": []
          }
        ],
        "summary": "One assignment's state, blinded until reveal",
        "tags": [
          "expert-review"
        ],
        "x-pgx-required-roles": [
          "EXPERT_REVIEWER"
        ]
      }
    },
...
  },
...
}